fix(sft): preserve unweighted token normalization

+++ src/prime_rl/trainer/sft/loss.py
@@ -11,7 +11,7 @@ def reduce_token_loss(
     if token_loss.shape != loss_mask.shape:
         raise ValueError(f"token_loss and loss_mask shapes differ: {token_loss.shape} != {loss_mask.shape}")
     if loss_weight is None:
-        return token_loss[loss_mask].sum(), loss_mask.sum(dtype=torch.float32)
+        return token_loss[loss_mask].sum(), loss_mask.sum(dtype=torch.int64)
     if loss_weight.shape != token_loss.shape:
         raise ValueError(f"loss_weight and token_loss shapes differ: {loss_weight.shape} != {token_loss.shape}")
